added sanity in fileservice uploads

+++ services/file-service/apps/files/services.py
@@ -2,32 +2,36 @@ from .models import UploadedFile
 from .exceptions import FileNotFoundException, FileUploadError, UnsupportedFileFormatError
 import cloudinary
 import cloudinary.uploader
-from django.conf import settings
-
-cloudinary.config(
-    cloud_name=settings.CLOUDINARY_CLOUD_NAME,
-    api_key=settings.CLOUDINARY_API_KEY,
-    api_secret=settings.CLOUDINARY_API_SECRET
-)
-
 class FileService:
 
-    def cloudinary_upload(self, file):
+    def cloudinary_upload(self, file, resource_type, resource_id, slot):
         if file.content_type not in ['image/jpeg', 'image/png', 'image/webp', 'image/gif']:
             raise UnsupportedFileFormatError("Unsupported file format for Product Image. Allowed formats: JPEG, PNG, WEBP, GIF.")
-        
+
+        if resource_type and resource_id and slot:
+            public_id = f"nexuscommerce/{resource_type}/{resource_id}/{slot}"
+        else:
+            public_id = f"nexuscommerce/misc/{file.name}"
+
         try:
             result = cloudinary.uploader.upload(
                         file,
                         quality="auto",
+                        public_id=public_id,
                         fetch_format="auto",
+                        overwrite=True
                     )
         except Exception as e:
+            print(f"Cloudinary error: {e}")
             raise FileUploadError("Error occurred while uploading file to Cloudinary.")
         return result.get('secure_url'), result.get('public_id')
-    
-    def upload_file(self, file, uploaded_by):
-        url , public_id = self.cloudinary_upload(file)
+
+    def upload_file(self, uploaded_by, **resource_info):
+        resource_type = resource_info.get('resource_type')
+        resource_id = resource_info.get('resource_id')
+        slot = resource_info.get('slot')
+        file = resource_info.get('file')
+        url, public_id = self.cloudinary_upload(file, resource_type, resource_id, slot)
 
         new_file = UploadedFile.objects.create(
             cloudinary_url=url,
@@ -51,7 +55,9 @@ class FileService:
         if not file:
             raise FileNotFoundException("File not found.")
         public_id = file.public_id
-        cloudinary.uploader.destroy(public_id)
+        result = cloudinary.uploader.destroy(public_id)
+        if result.get('result') not in ['ok', 'not found']:
+            raise FileUploadError("Error occurred while deleting file from Cloudinary.")
         file.delete()  
         return True
